+++ src/main/java/com/books/library/service/impl/AuthorServiceImple.java
@@ -1,7 +1,8 @@
-package com.books.library.service;
+package com.books.library.service.impl;
 
 import com.books.library.dto.Author;
 import com.books.library.repos.AuthorDao;
+import com.books.library.service.AuthorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,7 +10,7 @@ import java.util.List;
 
 
 @Service
-public class AuthorServiceImple implements AuthorService{
+public class AuthorServiceImple implements AuthorService {
     private final AuthorDao authorDao;
 
     @Autowired
